refactor: 설정 화면 컴포넌트 분리 #24
