+++ components/settings/SettingsHeader.tsx
@@ -1,0 +1,48 @@
+// src/components/settings/SettingsHeader.tsx
+import { Ionicons } from '@expo/vector-icons';
+import { router } from 'expo-router';
+import React from 'react';
+import { StyleSheet, Text, TouchableOpacity, View } from 'react-native';
+
+interface SettingsHeaderProps {
+    title: string;
+}
+
+export const SettingsHeader: React.FC<SettingsHeaderProps> = ({ title }) => {
+    const handleBack = () => {
+        router.back();
+    };
+
+    return (
+        <View style={styles.header}>
+            <TouchableOpacity style={styles.backButton} onPress={handleBack}>
+                <Ionicons name="chevron-back" size={24} color="#333" />
+            </TouchableOpacity>
+            <Text style={styles.headerTitle}>{title}</Text>
+            <View style={styles.placeholder} />
+        </View>
+    );
+};
+
+const styles = StyleSheet.create({
+    header: {
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        paddingHorizontal: 20,
+        paddingVertical: 15,
+        borderBottomWidth: 1,
+        borderBottomColor: '#f0f0f0',
+    },
+    backButton: {
+        padding: 5,
+    },
+    headerTitle: {
+        fontSize: 18,
+        fontWeight: 'bold',
+        color: '#333',
+    },
+    placeholder: {
+        width: 34,
+    },
+});
